+++ R/dynamite.R
@@ -131,21 +131,19 @@
 #' )
 #'
 #' if (requireNamespace("dplyr") && requireNamespace("ggplot2")) {
-#'   library(dplyr)
-#'   library(ggplot2)
 #    cf <- coef(fit) %>%
-#'   group_by(time, variable) %>%
-#'     summarise(
+#'   dplyr::group_by(time, variable) %>%
+#'     dplyr::summarise(
 #'       mean = mean(value),
 #'       lwr = quantile(value, 0.025),
 #'       upr = quantile(value, 0.975)
 #'     )
 #'   cf %>%
-#'     ggplot(aes(time, mean)) +
-#'     theme_bw() +
-#'     geom_ribbon(aes(ymin = lwr, ymax = upr), alpha = 0.7) +
-#'     geom_line() +
-#'     facet_wrap(~variable, scales = "free_y")
+#'     ggplot2::ggplot(aes(time, mean)) +
+#'     ggplot2::theme_bw() +
+#'     ggplot2::geom_ribbon(aes(ymin = lwr, ymax = upr), alpha = 0.7) +
+#'     ggplot2::geom_line() +
+#'     ggplot2::facet_wrap(~variable, scales = "free_y")
 #'   }
 #' }
 #'
@@ -243,7 +241,7 @@ dynamite <- function(dformula, data, group = NULL, time,
       stan = stan_input,
       group_var = group,
       time_var = time,
-      priors = dplyr::bind_rows(stan_input$priors),
+      priors = data.table::setDF(data.table::rbindlist(stan_input$priors)),
       backend = backend,
       call = match.call()
     ),